Add confirmation dialog

+++ apps/client/src/app/components/admin-market-data/admin-market-data.component.ts
@@ -178,10 +178,20 @@ export class AdminMarketDataComponent
   }
 
   public onDeleteProfileData({ dataSource, symbol }: UniqueAsset) {
-    this.adminService
-      .deleteProfileData({ dataSource, symbol })
-      .pipe(takeUntil(this.unsubscribeSubject))
-      .subscribe(() => {});
+    const confirmation = confirm(
+      $localize`Do you really want to delete this asset profile?`
+    );
+
+    if (confirmation) {
+      this.adminService
+        .deleteProfileData({ dataSource, symbol })
+        .pipe(takeUntil(this.unsubscribeSubject))
+        .subscribe(() => {
+          setTimeout(() => {
+            window.location.reload();
+          }, 300);
+        });
+    }
   }
 
   public onGather7Days() {
